+++ src/referee/mod.rs
@@ -38,7 +38,7 @@ pub(crate) fn referee_initial_setup(
     their_reward_puzzle_hash: &PuzzleHash,
     their_reward_payout_signature: &Aggsig,
     reward_puzzle_hash: &PuzzleHash,
-    nonce: usize,
+    nonce: u64,
     agg_sig_me_additional_data: &Hash,
 ) -> Result<RefereeInitialSetup, Error> {
     let initial_move = GameMoveStateInfo {
@@ -120,7 +120,7 @@ impl Referee {
         their_reward_puzzle_hash: &PuzzleHash,
         their_reward_payout_signature: &Aggsig,
         reward_puzzle_hash: &PuzzleHash,
-        nonce: usize,
+        nonce: u64,
         agg_sig_me_additional_data: &Hash,
         state_number: usize,
     ) -> Result<(Self, PuzzleHash), Error> {
